% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-helpers.R
\name{clopper_pearson}
\alias{clopper_pearson}
\title{Clopper-Pearson confidence interval for binary data}
\usage{
clopper_pearson(x, n, conf_level = 0.95)
}
\arguments{
\item{x}{Number of successes}

\item{n}{Total number of trials}

\item{conf_level}{Confidence level}
}
\value{
Named numeric vector, with confidence level stored as an attribute
}
\description{
Clopper-Pearson confidence interval for binary data
}
\examples{
clopper_pearson(1, 10)

}
