% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-helpers.R
\name{logit}
\alias{logit}
\alias{invlogit}
\title{Logit and inverse logit functions}
\usage{
logit(x)

invlogit(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Numeric vector
}
\description{
Logit and inverse logit functions
}
\examples{
logit((1:9)/10)
invlogit(-3:3)
logit(invlogit(-3:3))
}
