% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-logistic-data-v06.R
\docType{data}
\name{lr_dat}
\alias{lr_dat}
\title{Sample simulated data for logistic regression exposure-response models with covariates.}
\format{
A data frame with columns:
\describe{
\item{id}{Identifier}
\item{dose}{Nominal dose, units not specified}
\item{exposure}{Exposure value, units and metric not specified}
\item{exposure_quartile}{Exposure quartile, with placebo group separate}
\item{response}{Continuous response value (units not specified)}
\item{sex}{Sex}
}
}
\usage{
lr_dat
}
\description{
Sample simulated data for logistic regression exposure-response models with covariates.
}
\details{
This simulated dataset is entirely synthetic
You can find the data generating code in the package source code
}
\examples{
lr_dat
}
\keyword{datasets}
