% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-logistic-v06.R
\name{lr_model}
\alias{lr_model}
\title{Fit a logistic regression function}
\usage{
lr_model(formula, data, ...)
}
\arguments{
\item{formula}{Model formula}

\item{data}{Data set}

\item{...}{Other arguments passed to \code{glm()}}
}
\value{
A glm object
}
\description{
Fit a logistic regression function
}
\examples{
mod <- lr_model(response ~ exposure, lr_data)
mod

}
