% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-logistic-v06.R
\name{lr_plot}
\alias{lr_plot}
\alias{lr_plot_add_base}
\alias{lr_plot_add_strips}
\alias{lr_plot_build}
\title{Builds an exposure-response plot for a logistic regression model}
\usage{
lr_plot(obs_data, prd_data, exposure, response, exp_bins, ...)

lr_plot_add_base(plt)

lr_plot_add_strips(plt, shade = NULL)

lr_plot_build(plt)
}
\arguments{
\item{obs_data}{Observed data}

\item{prd_data}{Prediction data}

\item{exposure}{Exposure variable (unquoted)}

\item{response}{Response variable (unquoted)}

\item{exp_bins}{Exposure bin variable (unquoted)}

\item{...}{Other arguments}
}
\value{
A plot
}
\description{
Builds an exposure-response plot for a logistic regression model
}
\examples{
# add example here
}
