% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr-scm.R
\name{lr_scm}
\alias{lr_scm}
\alias{lr_scm_forward}
\alias{lr_scm_backward}
\alias{lr_scm_history}
\title{Stepwise covariate modelling for logistic regression}
\usage{
lr_scm_forward(mod, candidates, threshold = 0.01, seed = NULL)

lr_scm_backward(mod, candidates, threshold = 0.001, seed = NULL)

lr_scm_history(mod)
}
\arguments{
\item{mod}{An erlr model object}

\item{candidates}{Character vector with list of candidate terms}

\item{threshold}{Threshold to test against}

\item{seed}{Optional seed to control order of term tests}
}
\value{
For \code{lr_scm_forward()} and \code{lr_scm_backward()}, the updated erlr model
is returned, with the SCM history log updated internally. For \code{lr_scm_history()},
a data frame is returned containing the SCM history log
}
\description{
Stepwise covariate modelling for logistic regression
}
\examples{
mod0 <- lr_model(response_1 ~ exposure_1, lr_data)
mod1 <- lr_scm_forward(mod0, candidates = c("sex", "dose"))
lr_scm_history(mod1)

mod2 <- lr_model(response_1 ~ exposure_1 + sex + dose, lr_data)
mod3 <- lr_scm_backward(mod2, candidates = c("sex", "dose"))
lr_scm_history(mod3)
}
