% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-logistic-v06.R
\name{lr_vpc_plot}
\alias{lr_vpc_plot}
\title{Plot VPC simulations for logistic regression models}
\usage{
lr_vpc_plot(object, sim, group_by, conf_level = 0.95)
}
\arguments{
\item{object}{Logistic regression model}

\item{sim}{VPC simulations}

\item{group_by}{Variable (unquoted) to stratify predictions}

\item{conf_level}{Confidence level}
}
\description{
Plot VPC simulations for logistic regression models
}
