% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr-vpc.R
\name{lr_vpc_sim}
\alias{lr_vpc_sim}
\title{VPC simulations for logistic regression models}
\usage{
lr_vpc_sim(object, nsim = 100, seed = NULL)
}
\arguments{
\item{object}{Logistic regression model}

\item{nsim}{Number of replicates}

\item{seed}{RNG state}
}
\value{
A data frame or tibble
}
\description{
VPC simulations for logistic regression models
}
\examples{
mod <- lr_model(ae2 ~ aucss + sex, lr_data)
sim <- lr_vpc_sim(mod)
sim

}
