% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-logistic-v07.R
\name{logit}
\alias{logit}
\alias{invlogit}
\title{Logit and inverse logit functions}
\usage{
logit(x)

invlogit(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Numeric vector
}
\description{
Logit and inverse logit functions
}
\examples{
logit(lr_data$exposure_1)
invlogit(lr_data$response_1)
}
