% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-logistic-v06.R
\name{logit}
\alias{logit}
\alias{invlogit}
\title{Logit and inverse logit functions}
\usage{
logit(x)

invlogit(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Numeric vector
}
\description{
Logit and inverse logit functions
}
\examples{
logit(lr_dat$exposure)
invlogit(lr_dat$response)
}
