% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr-data.R
\docType{data}
\name{lr_data}
\alias{lr_data}
\title{Sample simulated data for logistic regression exposure-response models with covariates}
\format{
A data frame with columns:
\describe{
\item{id}{Identifier}
\item{sex}{Sex}
\item{age}{Age}
\item{weight}{Weight}
\item{dose}{Nominal dose, units not specified}
\item{treatment}{Treatment}
\item{aucss}{AUCss}
\item{cmaxss}{Cmax,ss}
\item{ae1}{Binary response 1 value}
\item{ae2}{Binary response 2 value}
}
}
\usage{
lr_data
}
\description{
Sample simulated data for logistic regression exposure-response models with covariates
}
\details{
This simulated dataset is entirely synthetic
You can find the data generating code in the package source code
}
\examples{
lr_data
}
\keyword{datasets}
