% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-logistic-data-v07.R
\docType{data}
\name{lr_data}
\alias{lr_data}
\title{Sample simulated data for logistic regression exposure-response models with covariates.}
\format{
A data frame with columns:
\describe{
\item{id}{Identifier}
\item{dose}{Nominal dose, units not specified}
\item{exposure_1}{Exposure 1 value, units and metric not specified}
\item{quartile_1}{Exposure 1 quartile, with placebo group separate}
\item{response_1}{Binary response 1 value}
\item{response_2}{Binary response 2 value}
\item{sex}{Sex}
}
}
\usage{
lr_data
}
\description{
Sample simulated data for logistic regression exposure-response models with covariates.
}
\details{
This simulated dataset is entirely synthetic
You can find the data generating code in the package source code
}
\examples{
lr_data
}
\keyword{datasets}
