% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr-plot.R
\name{lr_plot}
\alias{lr_plot}
\alias{lr_plot_add_model}
\alias{lr_plot_add_quantiles}
\alias{lr_plot_add_strips}
\alias{lr_plot_add_boxplot}
\title{Builds an exposure-response plot for a logistic regression model}
\usage{
lr_plot(data, exposure, response, color_by = NULL, ...)

lr_plot_add_model(object, color_by = "inherit")

lr_plot_add_quantiles(
  object,
  color_by = "inherit",
  bins = 4,
  conf_level = 0.95
)

lr_plot_add_strips(
  object,
  color_by = "inherit",
  style = "jitter",
  panel = "both"
)

lr_plot_add_boxplot(object, boxes_by, color_by = "inherit")
}
\arguments{
\item{data}{Observed data}

\item{exposure}{Exposure variable (unquoted)}

\item{response}{Response variable (unquoted)}

\item{color_by}{Variable (unquoted) to assign colors}

\item{...}{Other arguments}

\item{object}{Partially constructed plot (has S3 class \code{erlr_plot})}

\item{bins}{Number of exposure bins (not counting placebo)}

\item{conf_level}{Confidence level for Clopper-Pearson intervals}

\item{style}{Character string: "jitter" (the default) or "dotplot"}

\item{panel}{Character string: "upper", "lower", or "both" (the default)}

\item{boxes_by}{Variable (unquoted) to use to stratify exposure boxplots}
}
\value{
Plot object of class \code{erlr_plot}
}
\description{
Builds an exposure-response plot for a logistic regression model
}
\examples{
lr_data |> 
  lr_plot(exposure_1, response_1) |> 
  lr_plot_add_model() |> 
  lr_plot_add_quantiles() |> 
  lr_plot_add_boxplot(quartile_1) |> 
  print()

lr_data |> 
  lr_plot(exposure_1, response_1, sex) |> 
  lr_plot_add_model() |> 
  lr_plot_add_quantiles() |> 
  lr_plot_add_strips() |> 
  lr_plot_add_boxplot(quartile_1) |> 
  print()  

lr_data[1:70,] |> 
  lr_plot(exposure_1, response_1) |> 
  lr_plot_add_model() |> 
  lr_plot_add_quantiles(bins = 6) |> 
  lr_plot_add_strips(sex, style = "dotplot") |> 
  lr_plot_add_boxplot(quartile_1) |> 
  lr_plot_add_boxplot(sex) |> 
  print(box_height = 2)

}
