% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr-plot-api.R
\name{lr_plot}
\alias{lr_plot}
\alias{lr_plot_style}
\alias{lr_plot_show_model}
\alias{lr_plot_show_quantiles}
\alias{lr_plot_show_datastrip}
\alias{lr_plot_show_groups}
\alias{lr_plot_build}
\title{Builds an exposure-response plot for a logistic regression model}
\usage{
lr_plot(data, exposure, response, color_by = NULL)

lr_plot_style(object, labels)

lr_plot_show_model(object, use_color = NULL, conf_level = 0.95)

lr_plot_show_quantiles(object, use_color = NULL, bins = 4, conf_level = 0.95)

lr_plot_show_datastrip(
  object,
  use_color = NULL,
  style = "jitter",
  panel = "both"
)

lr_plot_show_groups(object, group_by, use_color = NULL)

lr_plot_build(object)
}
\arguments{
\item{data}{Observed data}

\item{exposure}{Exposure variable (one variable, unquoted)}

\item{response}{Response variable (one variable, unquoted)}

\item{color_by}{Stratification variable used for color and fill (one variable, unquoted)}

\item{object}{Partially constructed plot (has S3 class \code{erlr_plot})}

\item{labels}{Named list of labels}

\item{use_color}{Logical, indicating whether this component should keep the color stratification}

\item{conf_level}{Confidence level}

\item{bins}{Number of exposure bins (not counting placebo)}

\item{style}{Character string: "jitter" (the default) or "dotplot"}

\item{panel}{Character string: "upper", "lower", or "both" (the default)}

\item{group_by}{Stratification variables to define groups for boxplots (a tidyselection of variables)}
}
\value{
Plot object of class \code{erlr_plot}
}
\description{
Builds an exposure-response plot for a logistic regression model
}
\examples{
lr_data |> 
  lr_plot(exposure_1, response_1) |> 
  lr_plot_show_model() |> 
  lr_plot_show_quantiles() |> 
  lr_plot_show_groups(quartile_1) |> 
  plot()

lr_data |> 
  lr_plot(exposure_1, response_1, sex) |> 
  lr_plot_show_model() |> 
  lr_plot_show_quantiles() |> 
  lr_plot_show_datastrip() |> 
  lr_plot_show_groups(quartile_1) |> 
  plot()  

lr_data |> 
  lr_plot(exposure_1, response_1, color_by = sex) |> 
  lr_plot_show_model(use_color = FALSE) |> 
  lr_plot_show_quantiles(bins = 3) |> 
  lr_plot_show_datastrip() |> 
  lr_plot_show_groups(group_by = c(quartile_1, dose), use_color = FALSE) |> 
  plot()

}
