% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-logistic-v06.R
\name{lr_plot}
\alias{lr_plot}
\alias{lr_plot_add_base}
\alias{lr_plot_add_strips}
\alias{print.erlr_plot}
\title{Builds an exposure-response plot for a logistic regression model}
\usage{
lr_plot(data, exposure, response, ...)

lr_plot_add_base(object, bins = 4)

lr_plot_add_strips(object, color = NULL)

\method{print}{erlr_plot}(x, ...)
}
\arguments{
\item{data}{Observed data}

\item{exposure}{Exposure variable (unquoted)}

\item{response}{Response variable (unquoted)}

\item{...}{Other arguments}

\item{object}{Partially constructed plot (has S3 class \code{erlr_plot})}

\item{bins}{Number of exposure bins (not counting placebo)}

\item{color}{Variable (unquoted) to assign colors to strip plot dots}
}
\value{
Plot object of class \code{erlr_plot}
}
\description{
Builds an exposure-response plot for a logistic regression model
}
\examples{
# add example here
}
