% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr-plot.R
\name{lr_plot}
\alias{lr_plot}
\alias{lr_plot_add_quantiles}
\alias{lr_plot_add_dotplot_strips}
\alias{lr_plot_add_jitter_strips}
\alias{lr_plot_add_boxplot}
\title{Builds an exposure-response plot for a logistic regression model}
\usage{
lr_plot(data, exposure, response, ...)

lr_plot_add_quantiles(object, bins = 4, conf_level = 0.95)

lr_plot_add_dotplot_strips(object, color_by = NULL)

lr_plot_add_jitter_strips(object, color_by = NULL)

lr_plot_add_boxplot(object, group_by)
}
\arguments{
\item{data}{Observed data}

\item{exposure}{Exposure variable (unquoted)}

\item{response}{Response variable (unquoted)}

\item{...}{Other arguments}

\item{object}{Partially constructed plot (has S3 class \code{erlr_plot})}

\item{bins}{Number of exposure bins (not counting placebo)}

\item{conf_level}{Confidence level for Clopper-Pearson intervals}

\item{color_by}{Variable (unquoted) to assign colors to strip plot dots}

\item{group_by}{Variable (unquoted) to use to stratify exposure boxplots}
}
\value{
Plot object of class \code{erlr_plot}
}
\description{
Builds an exposure-response plot for a logistic regression model
}
\examples{
lr_data |> 
  lr_plot(exposure_1, response_1) |> 
  lr_plot_add_quantiles(bins = 4) |> 
  lr_plot_add_boxplot(group_by = quartile_1) |> 
  print()

lr_data |> 
  lr_plot(exposure_1, response_1) |> 
  lr_plot_add_quantiles(bins = 4) |> 
  lr_plot_add_jitter_strips(color_by = sex) |> 
  lr_plot_add_boxplot(group_by = quartile_1) |> 
  print()  

lr_data[1:70,] |> 
  lr_plot(exposure_1, response_1) |> 
  lr_plot_add_quantiles(bins = 4) |> 
  lr_plot_add_dotplot_strips(color_by = sex) |> 
  lr_plot_add_boxplot(group_by = quartile_1) |> 
  lr_plot_add_boxplot(group_by = sex) |> 
  print(box_height = 2)

}
