% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr-core.R
\name{lr_predict}
\alias{lr_predict}
\title{Predictions and confidence intervals for logistic regression}
\usage{
lr_predict(object, newdata = NULL, conf_level = 0.95)
}
\arguments{
\item{object}{A logistic regression model}

\item{newdata}{Data frame containing cases to be predicted}

\item{conf_level}{Confidence level for the intervals}
}
\value{
A tibble
}
\description{
Predictions and confidence intervals for logistic regression
}
\examples{
mod <- lr_model(ae1 ~ aucss, lr_data)
prd <- lr_predict(mod, lr_data)
prd

}
