% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-logistic-v07.R
\name{lr_vpc_plot}
\alias{lr_vpc_plot}
\title{Plot VPC simulations for logistic regression models}
\usage{
lr_vpc_plot(object, sim, group_by, conf_level = 0.95)
}
\arguments{
\item{object}{Logistic regression model}

\item{sim}{VPC simulations}

\item{group_by}{Variable (unquoted) to stratify predictions}

\item{conf_level}{Confidence level}
}
\value{
A ggplot2 object
}
\description{
Plot VPC simulations for logistic regression models
}
\examples{
mod <- lr_model(response_1 ~ exposure_1 + sex, lr_data)
sim <- lr_vpc_sim(mod)
lr_vpc_plot(mod, sim, group_by = exposure_1)
lr_vpc_plot(mod, sim, group_by = sex)

}
